Separate borrowed code into separate file, add license
